Record team stats
